Modifica para contratos
