+++ src/contratos/consumidores.py
