+++ src/contratos/utils.py
